requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.3
pymongo==4.4.0
pthon-dotenv==1.0.0

